api-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
requests==2.32.3
python-multipart==0.0.20
google-generativeai==0.8.4
aiofiles==24.1.0
python-dotenv==1.0.1
firebase-admin==6.2.0
